third_party/WebKit/LayoutTests/platform/mac/external/wpt/XMLHttpRequest/anonymous-mode-unsupported-expected.txt

This is a testharness.js-based test.
FAIL XMLHttpRequest: anonymous mode unsupported assert_equals: The deprecated anonymous:true should be ignored, cookie sent anyway expected "Cookie: test=anonymous-mode-unsupported\n" but got "{\"error\": {\"message\": \"Traceback (most recent call last):\\n  File \\"/b/c/b/mac_layout/src/third_party/WebKit/Tools/Scripts/webkitpy/thirdparty/wpt/wpt/tools/wptserve/wptserve/handlers.py\\", line 246, in __call__\n    rv = self.func(request, response)\n  File \\"/b/c/b/mac_layout/src/third_party/WebKit/LayoutTests/external/wpt/XMLHttpRequest/resources/inspect-headers.py\\", line 31, in main\n    result = get_response(request.raw_headers, filter_value, filter_name)\nAttributeError: 'Request' object has no attribute 'raw_headers'\n\", \"code\": 500}}"
Harness: the test ran to completion.
